RT #42539: From Brian Phillips:
When ignoring modules that have been use'd because they're included in
distributions that have already been listed as a dependency, the loop is
using the wrong array ref (loopking of @$skip instead of @$modules).

As an example, this can be reproduced by running the test against a
distribution that lists "Moose" as a dependency but has a "use
Moose::Util::TypeConstraints" somewhere in the code.  Obviously, this
module is included in the "Moose" distribution and the code itself is
"expanding" the Moose distribution to list all the modules included but
it's simply not looping over the right list of modules.